refactor: remove redundant (and sometimes incorrect) component suffix from filenames
