+++ src/actor-sheet/actor-sheet-character.html
@@ -8,42 +8,42 @@
         height="180"
         width="180"
       />
-      {{> "systems/cryptomancer/shared/components/text-input/text-input.component.html" label="Race"
+      {{> "systems/cryptomancer/shared/components/text-input/text-input.html" label="Race"
       value=data.data.biography.race name="data.biography.race" placeholder="Race"}}
     </div>
     <div class="basic-bio">
       <div>
-        {{> "systems/cryptomancer/shared/components/text-input/text-input.component.html" label="Common Name"
+        {{> "systems/cryptomancer/shared/components/text-input/text-input.html" label="Common Name"
         value=actor.name name="name" placeholder="Common Name"}}
-        {{> "systems/cryptomancer/shared/components/text-input/text-input.component.html" label="True Name"
+        {{> "systems/cryptomancer/shared/components/text-input/text-input.html" label="True Name"
         value=data.data.biography.trueName name="data.biography.trueName" placeholder="True Name"}}
-        {{> "systems/cryptomancer/shared/components/text-input/text-input.component.html" label="Pronouns"
+        {{> "systems/cryptomancer/shared/components/text-input/text-input.html" label="Pronouns"
         value=data.data.biography.pronouns name="data.biography.pronouns" placeholder="Pronouns"}}
       </div>
       <div class="defenses">
         <div class="defense">
-          {{> "systems/cryptomancer/actor-sheet/components/defense.component.html"
+          {{> "systems/cryptomancer/actor-sheet/components/defense.html"
           label="Dodge"
           value=data.data.core.speed.value
           desc="vs melee/missile"
           }}
         </div>
         <div class="defense">
-          {{> "systems/cryptomancer/actor-sheet/components/defense.component.html"
+          {{> "systems/cryptomancer/actor-sheet/components/defense.html"
           label="Parry"
           value=data.data.core.power.value
           desc="vs melee"
           }}
         </div>
         <div class="defense">
-          {{> "systems/cryptomancer/actor-sheet/components/defense.component.html"
+          {{> "systems/cryptomancer/actor-sheet/components/defense.html"
           label="Take Cover"
           value=data.data.core.wits.value
           desc="vs missile"
           }}
         </div>
         <div class="defense">
-          {{> "systems/cryptomancer/actor-sheet/components/defense.component.html"
+          {{> "systems/cryptomancer/actor-sheet/components/defense.html"
           label="Resist"
           value=data.data.core.resolve.value
           desc="vs magic"
@@ -65,62 +65,62 @@
     <div class="tab features" data-group="primary" data-tab="features">
       <div class="cores">
         <div class="core">
-          {{> "systems/cryptomancer/actor-sheet/components/core-triad.component.html"
+          {{> "systems/cryptomancer/actor-sheet/components/core-triad.html"
           core=data.data.core.speed
           attrL=data.data.core.speed.attributes.agility
           attrR=data.data.core.speed.attributes.dexterity
           }}
           <div class="skills">
-            {{> "systems/cryptomancer/actor-sheet/components/skill-list.component.html"
+            {{> "systems/cryptomancer/actor-sheet/components/skill-list.html"
             attr=data.data.core.speed.attributes.agility
             }}
-            {{> "systems/cryptomancer/actor-sheet/components/skill-list.component.html"
+            {{> "systems/cryptomancer/actor-sheet/components/skill-list.html"
             attr=data.data.core.speed.attributes.dexterity
             }}
           </div>
         </div>
         <div class="core">
-          {{> "systems/cryptomancer/actor-sheet/components/core-triad.component.html"
+          {{> "systems/cryptomancer/actor-sheet/components/core-triad.html"
           core=data.data.core.power
           attrL=data.data.core.power.attributes.strength
           attrR=data.data.core.power.attributes.endurance
           }}
           <div class="skills">
-            {{> "systems/cryptomancer/actor-sheet/components/skill-list.component.html"
+            {{> "systems/cryptomancer/actor-sheet/components/skill-list.html"
             attr=data.data.core.power.attributes.strength
             }}
-            {{> "systems/cryptomancer/actor-sheet/components/resource-skill.component.html"
+            {{> "systems/cryptomancer/actor-sheet/components/resource-skill.html"
             attr=data.data.core.power.attributes.endurance
             resource=data.data.healthPoints
             }}
           </div>
         </div>
         <div class="core">
-          {{> "systems/cryptomancer/actor-sheet/components/core-triad.component.html"
+          {{> "systems/cryptomancer/actor-sheet/components/core-triad.html"
           core=data.data.core.wits
           attrL=data.data.core.wits.attributes.knowledge
           attrR=data.data.core.wits.attributes.cunning
           }}
           <div class="skills">
-            {{> "systems/cryptomancer/actor-sheet/components/skill-list.component.html"
+            {{> "systems/cryptomancer/actor-sheet/components/skill-list.html"
             attr=data.data.core.wits.attributes.knowledge
             }}
-            {{> "systems/cryptomancer/actor-sheet/components/skill-list.component.html"
+            {{> "systems/cryptomancer/actor-sheet/components/skill-list.html"
             attr=data.data.core.wits.attributes.cunning
             }}
           </div>
         </div>
         <div class="core">
-          {{> "systems/cryptomancer/actor-sheet/components/core-triad.component.html"
+          {{> "systems/cryptomancer/actor-sheet/components/core-triad.html"
           core=data.data.core.resolve
           attrL=data.data.core.resolve.attributes.presence
           attrR=data.data.core.resolve.attributes.willpower
           }}
           <div class="skills">
-            {{> "systems/cryptomancer/actor-sheet/components/skill-list.component.html"
+            {{> "systems/cryptomancer/actor-sheet/components/skill-list.html"
             attr=data.data.core.resolve.attributes.presence
             }}
-            {{> "systems/cryptomancer/actor-sheet/components/resource-skill.component.html"
+            {{> "systems/cryptomancer/actor-sheet/components/resource-skill.html"
             attr=data.data.core.power.attributes.presence
             resource=data.data.manaPoints
             }}
