+++ src/actor-sheet/actor-sheet.scss
@@ -1,7 +1,7 @@
-@import "./components/core-triad.component";
-@import "./components/defense.component";
-@import "./components/resource-skill.component";
-@import "./components/skill-list.component.scss";
+@import "./components/core-triad";
+@import "./components/defense";
+@import "./components/resource-skill";
+@import "./components/skill-list";
 
 .crypt-actor-sheet {
   header {
