+++ src/actor-sheet/actor-sheet.ts
@@ -18,7 +18,7 @@ export class CryptomancerActorSheet extends ActorSheet<
   static get defaultOptions() {
     return mergeObject(super.defaultOptions, {
       classes: ["cryptomancer", "sheet", "actor"],
-      template: "systems/cryptomancer/actor-sheet/actor-sheet.component.html",
+      template: "systems/cryptomancer/actor-sheet/actor-sheet.html",
       width: 630,
       height: 730,
       tabs: [
@@ -33,7 +33,7 @@ export class CryptomancerActorSheet extends ActorSheet<
 
   /** @override */
   get template() {
-    return `systems/cryptomancer/actor-sheet/actor-sheet-${this.actor.data.type}.component.html`;
+    return `systems/cryptomancer/actor-sheet/actor-sheet-${this.actor.data.type}.html`;
   }
 
   /* -------------------------------------------- */
