+++ src/actor-sheet/components/core-triad.scss
