+++ src/actor-sheet/components/defense.html
@@ -5,7 +5,7 @@
 -->
 <div class="crypt-defense">
   <div class="label">{{ label }}</div>
-  {{> "systems/cryptomancer/shared/components/core-input/core-input.component.html"
+  {{> "systems/cryptomancer/shared/components/core-input/core-input.html"
   value=value
   readonly=true
   size="small"
