+++ src/actor-sheet/components/resource-skill.scss
