+++ src/actor-sheet/components/skill-list.scss
