+++ src/cryptomancer.scss
@@ -30,6 +30,6 @@
   -moz-appearance: none;
 }
 
-@import "skill-check/skill-check.component";
-@import "actor-sheet/actor-sheet.component";
+@import "skill-check/skill-check";
+@import "actor-sheet/actor-sheet";
 @import "shared/components";
