+++ src/global.d.ts
@@ -1,6 +1,6 @@
 import { ActorData } from "@league-of-foundry-developers/foundry-vtt-types/src/foundry/common/data/module.mjs";
 import { Character } from "./shared/interfaces/cryptomancer";
-import { CryptomancerActor } from "./actor/actor.component";
+import { CryptomancerActor } from "./actor/actor";
 
 interface CharacterDataSource {
   type: "character";
