+++ src/shared/components/index.scss
@@ -1,3 +1,3 @@
-@import "./core-input/core-input.component";
-@import "./text-input/text-input.component";
-@import "./toggle-box/toggle-box.component";
+@import "./core-input/core-input";
+@import "./text-input/text-input";
+@import "./toggle-box/toggle-box";
