+++ src/shared/components/toggle-box/toggle-box.scss
