+++ src/skill-check/skill-check.scss
